refactor: RVItem 패키지 이동
