+++ app/src/main/java/com/woowa/banchan/ui/home/HomeRVTag.kt
@@ -8,27 +8,3 @@ const val LINEAR_VERTICAL = 4
 const val LINEAR_HORIZONTAL = 5
 const val GRID = 6
 
-sealed class RVItem {
-    abstract val id: Int
-
-    object Header : RVItem() {
-        override val id = 0
-    }
-
-    object SubHeader : RVItem() {
-        override val id = 1
-    }
-
-    object Footer : RVItem() {
-        override val id = 2
-    }
-
-    data class Item<T>(val item: T) : RVItem() { // 아이템 항목
-        override val id = 3
-    }
-
-    object Empty : RVItem() {
-        override val id = Integer.MAX_VALUE
-    }
-
-}
