+++ app/src/main/java/com/woowa/banchan/ui/home/adapter/soupside/SoupSideRVAdapter.kt
@@ -12,7 +12,7 @@ import com.woowa.banchan.databinding.ItemSoupSideHeaderBinding
 import com.woowa.banchan.domain.model.FoodItem
 import com.woowa.banchan.ui.home.HOME_HEADER
 import com.woowa.banchan.ui.home.HOME_ITEM
-import com.woowa.banchan.ui.home.RVItem
+import com.woowa.banchan.ui.common.recyclerview.RVItem
 import com.woowa.banchan.ui.home.SUB_HEADER
 import com.woowa.banchan.ui.home.adapter.HomeRVAdapter
 import com.woowa.banchan.ui.home.adapter.soupside.viewholder.SoupSideHeaderViewHolder
