+++ app/src/main/java/com/woowa/banchan/ui/home/main/adapter/MainRVAdapter.kt
@@ -10,6 +10,7 @@ import com.woowa.banchan.R
 import com.woowa.banchan.databinding.ItemHomeHeaderBinding
 import com.woowa.banchan.databinding.ItemMainHeaderBinding
 import com.woowa.banchan.databinding.ItemRecyclerviewBinding
+import com.woowa.banchan.ui.common.recyclerview.RVItem
 import com.woowa.banchan.ui.home.*
 import com.woowa.banchan.ui.home.adapter.HomeRVAdapter
 import com.woowa.banchan.ui.home.adapter.viewholder.HomeHeaderViewHolder
